+++ demo/citibike/V1.2__load_tables_from_s3.sql
@@ -1,5 +1,5 @@
 -- Set the database and schema context
-USE SCHEMA SNOWCHANGE_DEMO.PUBLIC;
+USE SCHEMA SCHEMACHANGE_DEMO.PUBLIC;
 
 -- Load the trips data
 COPY INTO TRIPS FROM @TRIPS