+++ CHANGELOG.md
@@ -3,10 +3,18 @@ All notable changes to this project will be documented in this file.
 
 *The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).*
 
+## [2.9.0] - 2021-04-02
+
+### Changed
+- Renamed snowchange to schemachange
+- Please be advised that the new default path for the change history table is now `METADATA.SCHEMACHANGE.CHANGE_HISTORY` (the schema name has been changed from `SNOWCHANGE` to `SCHEMACHANGE`). The easiest thing to do is rename the metadata schema in Snowflake. The other option is to leave it as is and use the `--change-history-table` (or `-c`) parameter to override the default.
+
+
 ## [2.8.0] - 2021-02-09
 
 ### Added
 - Added a new optional parameter `--dry-run` to preview scripts that will be applied by snowchange.
+
 
 ## [2.7.0] - 2021-01-22